requirements.txt
gradio
requests
pillow
pytesseract
chromadb
sentence-transformers
pypdf
faiss-cpu
numpy
transformers
torch
